#3832 Feedback Thomas start script
