+++ package.json
@@ -16,7 +16,7 @@
     "icons": "tsx scripts/icons",
     "clear-node-modules": "rimraf node_modules angular-workspace/node_modules angular-workspace/projects/component-library/node_modules packages/core/node_modules packages/dso-toolkit/node_modules packages/react/node_modules storybook/node_modules",
     "website": "pnpm --filter website start",
-    "start": "tsx scripts/start.ts"
+    "start": "tsx scripts/start.mts"
   },
   "dependencies": {
     "@azure/storage-file-datalake": "^12.31.0",
